// src/app/Models/GetPageByEmpClass.ts
// Generated by https://quicktype.io

export interface GetPageByEmpClass {
    Code:        string;
    Name:        null;
    Url:         null | string;
    Title:       string;
    Description: null;
    ParentCode:  string;
    Order:       number;
    site:        site[];
}
// Generated by https://quicktype.io

export interface site {
    Code:        string;
    Name:        null;
    Url:         string;
    Title:       string;
    Description: null;
    ParentCode:  string;
    Order:       number;
    site:        any[];
}