// src/app/Models/GetBaseByAuthByEmpIDDataClass.ts
// Generated by https://quicktype.io

export interface GetBaseByAuthByEmpIDDataClass {
    IsAdmin: boolean;
    Dept:    Dept[];
    Base:    Base[];
}

export interface Base {
    EmpID:        string;
    DeptID:       number;
    DisplayHcode: boolean;
    App7Day:      boolean;
}

export interface Dept {
    DeptID:       number;
    DeptCode:     any;
    Path:         string;
    DisplayHcode: boolean;
    App7Day:      boolean;
}